fix: Ajuste menores
