+++ swagger/swagger.json
@@ -31,10 +31,13 @@
         ],
         "responses": {
           "200": {
-            "description": ""
+            "description": "Retorna detalle de un contrato-general."
+          },
+          "400": {
+            "description": "Error en la solicitud."
           }
         },
-        "summary": "Retorna información de contrato-general por ID",
+        "summary": "Retorna información de contrato-general por ID.",
         "tags": [
           "ContratoGeneral"
         ]
